requirements.txt
scrapling[fetchers]
google-genai
aiomysql
python-dotenv
aiohttp
